# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEEZER CLONE API",
        "description": "Descripción de la API de Ejemplo",
        "version": "1.0"
    },
    "paths": {
        "/api/album/{id}": {
            "get": {
                "tags": [
                    "Album"
                ],
                "summary": "Get album by ID",
                "operationId": "80910e4888da67bc521b68ecf13ebb65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Album ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album found"
                    },
                    "404": {
                        "description": "Album not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/artist/{id}": {
            "get": {
                "tags": [
                    "Artist"
                ],
                "summary": "Get artist by ID",
                "operationId": "f68bc8b51d728f005079874517a66764",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Artist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist found"
                    },
                    "404": {
                        "description": "Artist not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User's email",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "User's name",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    },
                                    "country": {
                                        "description": "User's country",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User's email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/search": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search for albums or artists",
                "operationId": "2754ee3c6371023cefd5b35cf65e9e07",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    }
}